Main file
Given a map N x N, each cell of the map represents the height in particular point. The task was to calculate minimum sum of all height changes on the way from point A to point B given on the map. To accomplish this task was written modified Dijkstra algorithm and vector-based priority queue.